test: 테스트 소나큐브 이슈 적용

+++ src/main/java/com/plzgraduate/myongjigraduatebe/graduation/application/service/CalculateGraduationService.java
@@ -69,7 +69,7 @@ class CalculateGraduationService implements CalculateGraduationUseCase {
 		GraduationResult graduationResult,
 		TakenLectureInventory originalTakenLectureInventory
 	) {
-		double actualTakenCredit = originalTakenLectureInventory.calculateTotalCredit()
+		double actualTakenCredit = (double) originalTakenLectureInventory.calculateTotalCredit()
 			+ user.getTransferCredit().getNormalCulture()
 			+ user.getTransferCredit().getMajorLecture()
 			+ user.getTransferCredit().getFreeElective()
